2013-07-19  Sharath Kamath  <s.kamath@samsung.com>
        WebCLPlatform::getDevices(DEVICE_TYPE_ALL) is not working properly
        Issue #220

        Reviewed by Antonio.

        cl_platform_getDevices.html in
        https://github.com/SRA-SiliconValley/webkit-webcl-conformance.

        Changing deviceType from int to unsigned long.
        To match with WebCLInputValidator class and OpenCL values.

        * Modules/webcl/WebCLPlatform.cpp:
        (WebCore::WebCLPlatform::getDevices):
        * Modules/webcl/WebCLPlatform.h:
        (WebCLPlatform):
        * Modules/webcl/WebCLPlatform.idl:
        (WebCLPlatform):

+++ Source/WebCore/Modules/webcl/WebCLPlatform.cpp
@@ -75,7 +75,7 @@ WebCLGetInfo WebCLPlatform::getInfo(int platform_info, ExceptionCode& ec)
     return WebCLGetInfo();
 }
 
-Vector<RefPtr<WebCLDevice> > WebCLPlatform::getDevices(int deviceType, ExceptionCode& ec) const
+Vector<RefPtr<WebCLDevice> > WebCLPlatform::getDevices(unsigned long deviceType, ExceptionCode& ec) const
 {
     Vector<RefPtr<WebCLDevice> > devices;
     if (!platformObject()) {
